+++ app/serializers/paginated_array_serializer.rb
@@ -1,12 +1,14 @@
-class PaginatedArraySerializer < ActiveModel::ArraySerializer
-  def initialize(object, options = {})
-    super(object, options)
+class PaginatedArraySerializer < ActiveModel::Serializer::ArraySerializer
+  def initialize(objects, options = {})
+    options[:meta] ||= {}
+    options[:meta].merge!(
+      pagination: {
+        total: objects.total_count,
+        per_page: objects.limit_value,
+        page: objects.current_page
+      }
+    )
 
-    @options[meta_key] ||= {}
-    @options[meta_key][:pagination] = {
-      total: object.total_count,
-      per_page: object.limit_value,
-      page: object.current_page
-    }
+    super(objects, options)
   end
 end