+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Unreleased
 
-
+* Upgrade Active Model Serializer to 0.10.0.rc2
+  ([#182](https://github.com/fs/rails-base-api/pull/182))
 * Rename bin/bootstrap to bin/setup
-  ([#180](https://github.com/fs/rails-base-api/pull/180)))
+  ([#180](https://github.com/fs/rails-base-api/pull/180))
 * Replace zeus with [Spring](https://github.com/rails/spring) for fast Rails actions via pre-loading
 * Provide full resource for all requests
   ([#176](https://github.com/fs/rails-base-api/pull/176))